+++ src/app/(guest)/sharing/[slug]/page.tsx
@@ -1,0 +1,10 @@
+type Params = Promise<{ slug: string }>;
+
+export default async function PublicSharingPage({
+  params,
+}: {
+  params: Params;
+}) {
+  const { slug } = await params;
+  return <div>Hello from PublicSharing {slug}</div>;
+}
