+++ src/shared/server/auth.ts
